requirements.txt
imagehash==4.0
keras==2.3.1
opencv-python==4.1.2.30
numpy==1.18.1
pillow==7.0.0
pyscreenshot==1.0
pandas==1.0.3
scikit-learn==0.22.1
tensorflow==2.1.0
